refactor(#55): Position 은 dto가 아닌 하나의 클래스로 생각돼 common에 위치
