+++ server/src/main/java/com/hexacore/tayo/car/dto/UpdateCarRequestDto.java
@@ -1,5 +1,6 @@
 package com.hexacore.tayo.car.dto;
 
+import com.hexacore.tayo.common.Position;
 import lombok.*;
 import org.springframework.web.multipart.MultipartFile;
 import java.util.List;
@@ -10,7 +11,7 @@ public class UpdateCarRequestDto {
 
     private Integer feePerHour;
     private String address;
-    private CreatePositionRequestDto position;
+    private Position position;
     private String description;
     private List<MultipartFile> imageFiles;
     private List<Integer> imageIndexes;
